Create delete_ellipsoid_shape.py
In this script, the delete_atoms function takes as input the name of the structure data file and the semi-axes lengths a, b, and c of the ellipsoid. The function opens the file, reads the lines, and loops through the atoms to identify the ones that are outside of the ellipsoid. The indices of the atoms to be deleted are stored in a list and returned by the function.

The example usage of the function calculates the list of atoms to be deleted for a structure data file named "structure.data" and semi-axes lengths of 1.0, 0.5, and 0.5, respectively. The list of indices is printed to the console.

Note: This script assumes that the structure data file is in a specific format and that the positions of the atoms are stored in the lines starting from the 9th line of the file. You may need to modify the script to match the specific format of your structure data file.