Create delete_shapeof_atoms_fromthe_structure.py
This code defines a function delete_atoms that takes the file_name, shape, radius_x, and radius_y as arguments and deletes the atoms that are part of the specified shape (either a circle or an ellipse) while keeping the rest of the atoms. The resulting structure is stored in a new file named new_structure.data.

The example usage calls the delete_atoms function with the argument file_name="structure.data", shape="ellipse", radius_x=2, and radius_y=1. This will delete the atoms that are part of an ellipse with semi-major axis of 2 and semi-minor axis of 1, while keeping the rest of the atoms.

def delete_atoms(file_name, shape, radius_x, radius_y):
    """
    This function deletes the atoms that are part of a specified shape while keeping the rest of the atoms.

    Parameters:
    file_name (str): The name of the file that contains the structure information.
    shape (str): The shape of the region where the atoms will be deleted. The options are "circle" or "ellipse".
    radius_x (float): The x-radius of the shape.
    radius_y (float): The y-radius of the shape.

    Returns:
    None

    Output:
    new_structure.data: A new file that contains the resulting structure after deleting the atoms.
    """